暂时的

+++ src/main/java/com/nk/dao/MyfileMapper.java
@@ -17,5 +17,5 @@ public interface MyfileMapper {
 
     int updateByPrimaryKey(Myfile record);
 
-    List<Myfile> selectByFunId(Integer id);
+    List<Myfile> selectByFunId(Long id);
 }
